Update DB2Grammar.php
Changed compile over to not include column names. This was throwing an error while chunking results from the db2. The comma in "select $column_names," was throwing a SQL syntax error. As far as I can tell, the only time this is used is when chunking results, so the column names shouldn't be included.